refactor: move archetype experiments to mlcast.config.archetype subpackage
Split Experiment dataclass (base.py) from experiment-specific config
functions. Move convgru_training_experiment to archetype/convgru.py
and ldcast_training_experiment to archetype/ldcast.py.
